Use default overloads when calling custom ops
If a node is created without specifying an overload,
A OpOverloadPacket is created, rather than an OpOverload.
This works in a GraphModule, but the OpOverloadPacket is not a valid
operator type in the _EXIREdgeDialectVerifier, which means
that Edge ExportedPrograms can't contain a GraphModule with
such ops.

In short, specifying using the default overload seems
to be the more correct way of creating a custom
operator.

Signed-off-by: Erik Lundell <erik.lundell@arm.com>
Change-Id: I3a1733c0ae88826d88b1e820eaacff765df7fbd2